third_party/blink/web_tests/fast/events/mouse-event-buttons-attribute-expected.txt

===== doubleClickAction with initial state=[] =====
1. targetDiv received mouseover buttons: 0
2. targetDiv received mouseenter buttons: 0
3. targetDiv received mousemove buttons: 0
4. targetDiv received mousedown buttons: 1
5. targetDiv received mouseup buttons: 0
6. targetDiv received click buttons: 0
7. targetDiv received mousedown buttons: 1
8. targetDiv received mouseup buttons: 0
9. targetDiv received click buttons: 0
10. targetDiv received dblclick buttons: 0
11. targetDiv received mouseout buttons: 0
12. targetDiv received mouseleave buttons: 0


===== clickAction with initial state=[rightButton] =====
1. targetDiv received mouseover buttons: 2
2. targetDiv received mouseenter buttons: 2
3. targetDiv received mousemove buttons: 2
4. targetDiv received mousedown buttons: 3
5. targetDiv received mouseup buttons: 2
6. targetDiv received click buttons: 2
7. targetDiv received mouseout buttons: 2
8. targetDiv received mouseleave buttons: 2


===== clickAction with initial state=[middleButton,rightButton] =====
1. targetDiv received mouseover buttons: 6
2. targetDiv received mouseenter buttons: 6
3. targetDiv received mousemove buttons: 6
4. targetDiv received mousedown buttons: 7
5. targetDiv received mouseup buttons: 6
6. targetDiv received click buttons: 6
7. targetDiv received mouseout buttons: 6
8. targetDiv received mouseleave buttons: 6


===== moveAction with initial state=[leftButton,middleButton,rightButton] =====
1. targetDiv received mouseover buttons: 7
2. targetDiv received mouseenter buttons: 7
3. targetDiv received mousemove buttons: 7
4. targetDiv received mouseout buttons: 7
5. targetDiv received mouseleave buttons: 7


===== rightClickAction with initial state=[] and showContextMenuOnMouseUp=false =====
1. targetDiv received mouseover buttons: 0
2. targetDiv received mouseenter buttons: 0
3. targetDiv received mousemove buttons: 0
4. targetDiv received mousedown buttons: 2
5. targetDiv received contextmenu buttons: 2
6. targetDiv received mouseup buttons: 0
7. targetDiv received mouseout buttons: 0
8. targetDiv received mouseleave buttons: 0


===== rightClickAction with initial state=[] and showContextMenuOnMouseUp=true =====
1. targetDiv received mouseover buttons: 0
2. targetDiv received mouseenter buttons: 0
3. targetDiv received mousemove buttons: 0
4. targetDiv received mousedown buttons: 2
5. targetDiv received mouseup buttons: 0
6. targetDiv received contextmenu buttons: 0
7. targetDiv received mouseout buttons: 0
8. targetDiv received mouseleave buttons: 0


===== wheelAction with initial state=[leftButton,rightButton] =====
1. targetDiv received mouseover buttons: 3
2. targetDiv received mouseenter buttons: 3
3. targetDiv received mousemove buttons: 3
4. targetDiv received mousewheel buttons: 3
5. targetDiv received mouseout buttons: 3
6. targetDiv received mouseleave buttons: 3


===== doubleTapAction with initial state=[] =====
1. targetDiv received mouseover buttons: 0
2. targetDiv received mouseenter buttons: 0
3. targetDiv received mousemove buttons: 0
4. targetDiv received mousedown buttons: 1
5. targetDiv received contextmenu buttons: 0
6. targetDiv received mouseup buttons: 0
7. targetDiv received click buttons: 0
8. targetDiv received mousemove buttons: 0
9. targetDiv received mousedown buttons: 1
10. targetDiv received mouseup buttons: 0
11. targetDiv received click buttons: 0
12. targetDiv received dblclick buttons: 0
13. targetDiv received mouseout buttons: 0
14. targetDiv received mouseleave buttons: 0


===== tapAction with initial state=[] =====
1. targetDiv received mouseover buttons: 0
2. targetDiv received mouseenter buttons: 0
3. targetDiv received mousemove buttons: 0
4. targetDiv received mousedown buttons: 1
5. targetDiv received mouseup buttons: 0
6. targetDiv received click buttons: 0
7. targetDiv received mouseout buttons: 0
8. targetDiv received mouseleave buttons: 0


===== longPressAction with initial state=[] and showContextMenuOnMouseUp=false =====
1. targetDiv received mouseover buttons: 0
2. targetDiv received mouseenter buttons: 0
3. targetDiv received mousemove buttons: 0
4. targetDiv received contextmenu buttons: 0
5. targetDiv received mouseout buttons: 0
6. targetDiv received mouseleave buttons: 0


===== longPressAction with initial state=[] and showContextMenuOnMouseUp=true =====


===== longTapAction with initial state=[] and showContextMenuOnMouseUp=false =====
1. targetDiv received mouseover buttons: 0
2. targetDiv received mouseenter buttons: 0
3. targetDiv received mousemove buttons: 0
4. targetDiv received contextmenu buttons: 0
5. targetDiv received mouseout buttons: 0
6. targetDiv received mouseleave buttons: 0


===== longTapAction with initial state=[] and showContextMenuOnMouseUp=true =====
1. targetDiv received mouseover buttons: 0
2. targetDiv received mouseenter buttons: 0
3. targetDiv received mousemove buttons: 0
4. targetDiv received contextmenu buttons: 0
5. targetDiv received mouseout buttons: 0
6. targetDiv received mouseleave buttons: 0


===== dragDropAction with initial state=[] =====
1. dragImg received mouseover buttons: 0
2. dragImg received mouseenter buttons: 0
3. dragImg received mousemove buttons: 0
4. dragImg received mousedown buttons: 1
5. dragImg received mouseout buttons: 1
6. dragImg received mouseleave buttons: 1
7. targetDiv received mouseover buttons: 1
8. targetDiv received mouseenter buttons: 1
9. targetDiv received mousemove buttons: 1
10. dragImg received dragstart buttons: 1
11. dragImg received drag buttons: 1
12. targetDiv received dragenter buttons: 1
13. dragImg received drag buttons: 1
14. dragImg received dragenter buttons: 1
15. targetDiv received dragleave buttons: 1
16. dragImg received drag buttons: 1
17. targetDiv received dragenter buttons: 1
18. dragImg received dragleave buttons: 1
19. targetDiv received dragover buttons: 0
20. dragImg received drag buttons: 0
21. targetDiv received dragleave buttons: 0
22. dragImg received dragend buttons: 0
23. targetDiv received mouseout buttons: 0
24. targetDiv received mouseleave buttons: 0


===== dragDropAction with initial state=[rightButton] =====
1. dragImg received mouseover buttons: 2
2. dragImg received mouseenter buttons: 2
3. dragImg received mousemove buttons: 2
4. dragImg received mousedown buttons: 3
5. dragImg received mouseout buttons: 3
6. dragImg received mouseleave buttons: 3
7. targetDiv received mouseover buttons: 3
8. targetDiv received mouseenter buttons: 3
9. targetDiv received mousemove buttons: 3
10. dragImg received dragstart buttons: 3
11. dragImg received drag buttons: 3
12. targetDiv received dragenter buttons: 3
13. dragImg received drag buttons: 3
14. dragImg received dragenter buttons: 3
15. targetDiv received dragleave buttons: 3
16. dragImg received drag buttons: 3
17. targetDiv received dragenter buttons: 3
18. dragImg received dragleave buttons: 3
19. targetDiv received dragover buttons: 2
20. dragImg received drag buttons: 2
21. targetDiv received dragleave buttons: 2
22. dragImg received dragend buttons: 0
23. targetDiv received mouseout buttons: 2
24. targetDiv received mouseleave buttons: 2


===== dragDropAction with initial state=[middleButton,rightButton] =====
1. dragImg received mouseover buttons: 6
2. dragImg received mouseenter buttons: 6
3. dragImg received mousemove buttons: 6
4. dragImg received mousedown buttons: 7
5. dragImg received mouseout buttons: 7
6. dragImg received mouseleave buttons: 7
7. targetDiv received mouseover buttons: 7
8. targetDiv received mouseenter buttons: 7
9. targetDiv received mousemove buttons: 7
10. dragImg received dragstart buttons: 7
11. dragImg received drag buttons: 7
12. targetDiv received dragenter buttons: 7
13. dragImg received drag buttons: 7
14. dragImg received dragenter buttons: 7
15. targetDiv received dragleave buttons: 7
16. dragImg received drag buttons: 7
17. targetDiv received dragenter buttons: 7
18. dragImg received dragleave buttons: 7
19. targetDiv received dragover buttons: 6
20. dragImg received drag buttons: 6
21. targetDiv received dragleave buttons: 6
22. dragImg received dragend buttons: 0
23. targetDiv received mouseout buttons: 6
24. targetDiv received mouseleave buttons: 6


PASS successfullyParsed is true

TEST COMPLETE
